[#26629] YSQL: Push down ef_search parameter to ybhnsw
Summary:
The GUC `hnsw.ef_search` in pgvector controls the expansion factor for HNSW searches, where higher values improve accuracy at the cost of increased latency. This change introduces a corresponding GUC, `ybhnsw.ef_search`, and propagates it to DocDB. Additional work is required in DocDB to utilize the propagated value.

**Upgrade/Rollback safety:**
This change adds new fields to `PgVectorReadOptionsPB` in common.proto that are unused in DocDB code as of this change. This does not suggest any upgrade/rollback issues.

Test Plan:
Jenkins

Manual testing by observing the docdb requests of ybhnsw queries to make sure the correct ef_search parameter is passed down

Reviewers: kramanathan

Reviewed By: kramanathan

Subscribers: yql

Differential Revision: https://phorge.dev.yugabyte.com/D42934