[#25562] CDC: Fix flag value conversion in update restart time code
Summary:
At the end of each GetConsistentChanges call we check whether restart time can be moved forward based on `last_received_restart_lsn` and the `last_shipped_lsn`. In the case where CDC is caught up, i.e when `last_received_restart_lsn == last_shipped_lsn` we update the restart time to the latest `vwal_safe_time` and write this restart time in the cdc_state table. However we don't want to write to the cdc_state table for every GetConsistentChanges call, so the flag `cdcsdk_update_restart_time_interval_secs` controls the frequency of these writes.

Due to a bug this limiting logic was not getting executed. This is because the value of the flag is in seconds whereas the variables used to track the last time when write to cdc_state table was performed maintain time in microseconds. The issue was fixed in D42872, this revision modifies an existing test to validate the correctness.
Jira: DB-14815

Test Plan: ./yb_build.sh --cxx-test integration-tests_cdcsdk_consumption_consistent_changes-test --gtest_filter CDCSDKConsumptionConsistentChangesTest.TestMovingRestartTimeForwardWhenNothingToStream

Reviewers: skumar, vkushwaha, sergei

Reviewed By: vkushwaha

Subscribers: ycdcxcluster

Tags: #jenkins-ready

Differential Revision: https://phorge.dev.yugabyte.com/D42947