[#26596] YSQL: Dont bump catalog version for temp table DDLs
Summary:
Do not bump catalog version for temp table/index DDLs Create/Drop/Alter.
This is because temporary relations are local to the node and do not need to be visible to other nodes. We set `is_altering_existing_data`, so that the local nodes tserver response cache is cleared for Alter and Drop since it will not be valid anymore.

Also, bumping the catalog version is not allowed during YSQL Major upgrade, which would cause the DDL to fail.

Added additional validation in `YsqlInitDBAndMajorUpgradeHandler::IsWriteToCatalogTableAllowed` to ensure the catalog version is not bumped in the middle of a YSQL upgrade.

- Updated `pg_ddl_atomicity_stress-test` to ignore "duplicate key value violates unique constraint". Before this change it would also occur with "Catalog Version Mismatch", but the extra check for temp table means the errors occur independently now.
- Updated `yb.orig.catalog_version` since now the catalog versions are not bumped for temp table DDls.
Jira: DB-15966

Test Plan:
ybd release --cxx-test integration-tests_ysql_major_upgrade_ddl_blocking-test --gtest_filter YsqlMajorUpgradeDdlBlockingTest.TestDdlsDuringUpgrade

PgCatalogVersionTest.NonIncrementingDDLMode

Reviewers: myang, telgersma, fizaa, dmitry

Reviewed By: myang, fizaa

Subscribers: svc_phabricator, dmitry, yql, ybase

Differential Revision: https://phorge.dev.yugabyte.com/D42866