[#25853] xCluster: move ground truth of replication role to system catalog
Summary:
Currently, for xCluster automatic mode, we install a DDL replication extension on both universes and change an extension-specific per-database variable to control what role the extension believes it is playing for that database.  (E.g., if it's supposed to be the source then it should be capturing DDL's.)

Unfortunately, this variable is only read by Postgres backends on startup so changing roles (e.g., during a switchover operation) does not affect existing Postgres connections.  :-(  This means among other things that if you start up replication while you have an existing Postgres connection, any DDL performed on that connection after the replication is set up will not be replicated.

To solve this problem, this diff switches to using a new mechanism: the replication role for each database is now kept in the system catalog in the xCluster config system catalog entry.

This information is distributed to the TServers via the master heartbeat response and stored in the TServer xcluster_context.  The extension before each DDL calls the local TServer via pggate to get the current role information for the current database of that Postgres connection.

Note that there is lag here -- the extension will not see the new information until after the heartbeat has occurred.  We are planning a additional mechanism to "wait until every TServer has received a heartbeat response" to handle this lag.  For now, the tests have sleeps and the xCluster replication cannot handle DDL's concurrently with set up and switch over.  (The latter is a known to do task and needs other work.)

In general, this diff is not intended to make other functional changes besides this mechanism switch; some small amount of per-files tidying has been included on the leave the code better than you found it principle.

Other details:
  * a new extension function, get_replication_role, is provided to return the current role the extension believes it is playing for the current connections database.
    * this is provided both for testing and for future debugging
  * a new extension procedure, TEST_override_replication_role, is provided to override the current role for the current connection
    * again this allows testing but also may help with future investigations
    * note this is only for the current connection unlike the previous mechanisms
  * usefully, if the extension is created on a database not under xCluster replication, it does nothing
    * this can occur if you restore a backup taken while xCluster replication was running

Note that there is a brief window after a TServer starts where it has not yet received a heartbeat response from master; if a Postgres connection is made and a DDL attempts to run, the DDL might fail due to role information unavailable.  It is possible we should add some invisible retries here to make that less likely.

Upgrade/rollback safety:

We are assuming (via documentation warnings for unstable releases), that no legal upgrade across this diff starts with automatic mode replication running.  (No previous stable release has automatic mode available; it is only available in 2.25.1, which can only legally be upgraded to 2.25.2.)    Accordingly, we do not have to worry about upgrade safety in the presence of automatic mode already being used/started during the upgrade.

This diff adds a new pggate RPC call:
```
~/code/yugabyte-db/src/yb/tserver/pg_client.proto:52:
  rpc GetXClusterRole(PgGetXClusterRoleRequestPB) returns (PgGetXClusterRoleResponsePB);
```
Pggate RPC calls don't matter for upgrades because they only go to the local TServer and are thus upgraded in lockstep.

This diff also adds a new field to the catalog namespace entries:
```
~/code/yugabyte-db/src/yb/common/common_types.proto:219:
// Namespaces (previously) under DB-scoped xCluster replication contain this information.
message XClusterNamespaceInfoPB {
    UNSPECIFIED = 0;         // This should never occur; not stored.
    UNAVAILABLE = 1;         // Used to denote that we cannot determine the role; not stored.

    NOT_AUTOMATIC_MODE = 2;  // No automatic mode replication is occurring for this namespace.

    AUTOMATIC_SOURCE = 3;
    AUTOMATIC_TARGET = 4;
  }
  required XClusterRole role = 1 [default = NOT_AUTOMATIC_MODE];
}

~/code/yugabyte-db/src/yb/master/catalog_entity_info.proto:442:
message SysXClusterConfigEntryPB {
  ...
  // Local NamespaceId -> xCluster info for that namespace.
  map<string, XClusterNamespaceInfoPB> xcluster_info_per_namespace = 3;
}
```
Given the reasonable default (empty map, a.k.a. no name spaces are under automatic replication), which matches the expectation that no automatic mode replication will be running as the upgrade starts, and the inability to use automatic mode during the upgrade, this cannot cause problems.

Additional information is added to the master heartbeat response:
```
~/code/yugabyte-db/src/yb/master/master_heartbeat.proto:266:
message TSHeartbeatResponsePB {
  ...
  // Local NamespaceId -> xCluster info for that namespace.
  map<string, XClusterNamespaceInfoPB> xcluster_info_per_namespace = 30;
}
```
Given the protobuf compatibility rules, this causes no problems.  Note
that in the absence of the DDL replication extension, nothing looks at
this information.

We are also changing the SQL objects created by the extension; we are not providing a migration from the previous version of the extension because again it should not be possible to upgrade from a version with the old extension.

What about upgrading 2024.2.2 to 2025.1.0 with Semi-Automatic xCluster running?  That causes no issues because there is no extension and in the absence of an extension none of these changes make a difference.

Fixes #25853
Jira: DB-15147

Test Plan:
New end-to-end test that existing connections have the role they use updated:
```
time ybd release --cxx-test xcluster_ddl_replication-test --gtest_filter '*.ExtensionRoleUpdating' --test-args '' >& /tmp/generic.mdl.log
```

New test to verify that we capture DDL's on existing connections after replication is set up:
```
time ybd release --cxx-test xcluster_ddl_replication-test --gtest_filter '*.CreateTableInExistingConnection' --test-args '' >& /tmp/generic.mdl.log
```

New regression test suite to verify that the new extension routines work correctly: postgres/yb-extensions/yb_xcluster_ddl_replication/sql/routines.sql
```
time ybd release --cxx-test xcluster_ddl_replication-test --test-args '' |& tee /tmp/generic.mdl.log | grep -P '^\[ '
```

Reviewers: xCluster, jhe

Reviewed By: jhe

Subscribers: svc_phabricator, hsunder, ybase, yql

Differential Revision: https://phorge.dev.yugabyte.com/D43178