[BACKPORT 2024.2][#25309] DocDB: Change resumption hybrid time for waiters in wait-queue resumed due to transaction promotion
Summary:
Original commit: 13a8a5a6798c13509fd217c02c95eb98da1461e2 / D42371
When waiters are resumed from the wait-queue due to transaction promotion (either of self, or any blocker), they are resumed with status ok and forced to re-enter the queue. We expect that the resumption hybrid time is valid when resuming with a status ok.

There are a few cases when the existing code resumes waiters with ht kMin on transaction promotion. This diff addresses one such case - If the participant processes the promotion message before the first transaction status response of `RunningTransaction`, we get the status ht as kMin.
Jira: DB-14508

Test Plan: AshGeoPartitionedTransactionTest.TestNonEmptyAshMetadata

Reviewers: esheng, rthallam

Reviewed By: esheng

Subscribers: ybase

Differential Revision: https://phorge.dev.yugabyte.com/D43001