[#26163] YSQL: Use PG's snapshot management for foreign key triggers
Summary:
Foreign key integrity in Pg is maintained using a set of 12 triggers. These are
`AFTER ROW` triggers that execute at the end of a statement. Some of these
triggers are also deferrable. In Pg, most of these fk triggers use a new latest
snapshot to perform operations. If this is not done, fk violations of the
following kind can occur:

```
Buggy YSQL behaviour:

create table pk_rel (h int primary key, v1 int unique);
create table fk_rel (h int primary key, v1 int references pk_rel(v1));
insert into pk_rel values (1, 1);

Session 1:                                           Session 2:

begin transaction isolation level repeatable read;
                                                     begin transaction isolation level repeatable read;
                                                     select count(*) from pk_rel; <---- Pick transaction snapshot
insert into fk_rel values (1, 1);
commit;
                                                     delete from pk_rel where h=1; <---- trigger operates on transaction snapshot instead of latest snapshot and hence doesn't see the concurrent insert
                                                     commit;
select * from pk_rel;
select * from fk_rel;
```

By default fks have an ON DELETE NO ACTION policy which means a DELETE
should fail if fk_rel has a row that references the row being deleted. So in
the above example, the DELETE fails with an error in Pg. To be able to
detect concurrently committed rows that could be referencing the to-be
DELETEd row, the "RI_FKey_noaction_del" system trigger uses a new
snapshot for querying fk_rel. Further, note that there can't be any other concurrently
inserted rows after the newly picked snapshot because they would require a
`for key share` lock on the pk_rel row that is already locked for deletion. In short,
the trigger snapshot can be one that is picked anytime after the pk_rel row is
provisionally deleted.

In YSQL, before this change, the `DELETE NO ACTION` trigger tries to read the
`fk_rel` table as of the transaction snapshot followed by acquiring a `FOR KEY SHARE`
lock on the read row. Since the row doesn't exist as of the transaction snapshot, they miss
to lock the concurrently inserted row which results in the above fk violation.

This change fixes that by using the latest snapshot for all fk triggers that require
doing so. All fk triggers call into `ri_PerformCheck` one or more times. The
`detectNewRows` parameter in this call is used to indicate that a new latest
snapshot is to be used.

YSQL was skipping the logic for `detectNewRows` because earlier we didn't
have a way to switch between snapshots in a transaction. 9124c053 added
support for switching snapshots in a Repeatable Read transaction.

Note that `detectNewRows` is true for all fk trigger queries except for
the 2 triggers that are executed when inserting or updating a row in the
referencing table (i.e., fk relation) AND they are not using partitioned tables.

NOTE: If the reference table (i.e., parent table) is partitioned, vanilla Pg
picks the latest snapshot for when performing a check to ensure existence of a
foreign key in the parent. So, the fk trigger is able to see concurrently
committed rows as well. However, the diff leaves the YSQL behaviour untouched
so as to keep the fk batching optimization intact (which was enabled as part of
feddbabf1c). Refer https://github.com/yugabyte/yugabyte-db/issues/28134 for more
details.
Jira: DB-15496

Test Plan: ./yb_build.sh release --java-test org.yb.pgsql.TestPgRegressWaitQueues

Reviewers: patnaik.balivada, dmitry

Reviewed By: dmitry

Subscribers: smishra, jason, yql

Differential Revision: https://phorge.dev.yugabyte.com/D46367