[BACKPORT 2024.2][#26659] CDC: Block drop table when the table is part of a publication
Summary:
#####Backport Description
Faced some merge conflicts in `tablecmds.c` due to PG 15 merge.

#####Original Description
When logical replication is running, if a user drops a table from the polling list slot then the slot being used is rendered useless. To prevent such scenarios this diff adds guardrails in the drop table code path. If the table being dropped is a part of publication then we fail the drop operation with appropriate error. To identify if the given table is a part of any publication we look at the `pg_publication_rel` catalog table.

One caveat here is that we do not block drop table when the table is included in publication(s) by the virtue of publication(s) being `ALL TABLES`. This is because a user cannot drop a table from an `ALL TABLES` publication.
Jira: DB-16036

Original commit: 02882cd2e83f679b77082846de537520be13e743 / D43026

Test Plan:
./yb_build.sh --cxx-test integration-tests_cdcsdk_consumption_consistent_changes-test --gtest_filter CDCSDKConsumptionConsistentChangesTest.TestBlockDropTableWhenPartOfPublication

The tests `TestPgRegressPublication` and `PgPartmanTest.TestTimeDailyNative` had to be modified since they dropped tables belonging to non-ALL TABLES publications.

Reviewers: skumar, vkushwaha, stiwary, aagrawal, jason

Reviewed By: vkushwaha

Subscribers: ycdcxcluster, yql

Tags: #jenkins-ready

Differential Revision: https://phorge.dev.yugabyte.com/D43078