[#26731] docdb: Fixed SIGSEGV inside ProcessSupervisor when can't restart process
Summary:
When ProcessSupervisor fails to restart exited process (for example due to hitting limit on number of threads), `ProcessSupervisor::process_wrapper_` remains `nullptr` and trying to wait for it inside `ProcessSupervisor::RunThread()` causes SIGSEGV.

Added check for nullptr and test that reproduces the issue. Test will be committed to the codebase as disabled because due to hitting limit on number of processes it will impact test server stability.
Jira: DB-16106

Test Plan: ProcessWrapperTest.HitThreadsLimit

Reviewers: arybochkin, #db-approvers

Reviewed By: arybochkin

Subscribers: ybase, yql

Tags: #jenkins-ready

Differential Revision: https://phorge.dev.yugabyte.com/D43098