[#26733] CDC: Reduce the defualt value of yb_walsender_poll_sleep_duration_empty_ms
Summary:
This diff reduces the value of the flag `yb_walsender_poll_sleep_duration_empty_ms` from 1 second to 10 mili seconds. This flag determines the interval walsender will sleep before calling the next GetConsistentChanges if the previous GetConsistentChanges returned an empty response.

This change is made because with a sleep of 1 second, when VWAL is started with large number of tablets, lot of time is taken to ship records when workload is very sparse. This happens because with a sparse workload, most of the tablet queues are empty and as a result most of the GetConsistentChanges calls return empty results to walsender. The walsender than sleeps for 1 second before calling next GetConsistentChanges. As a result if the first 'n-1' tablet queues are empty (contain only safepoints), to reach to the tablet 'n' with a valid record, walsender sleeps for 'n-1' seconds.

This scenario has often been encountered in customer test environments, where the workload is very sparse, giving an impression of slow replication speeds. To avoid this we are reducing the default value of this flag.
Jira: DB-16108

Test Plan: Jenkins: test regex: .*CDC.*|.*ReplicationSlot.*

Reviewers: vkushwaha, skumar

Reviewed By: vkushwaha

Subscribers: yql, ycdcxcluster

Tags: #jenkins-ready

Differential Revision: https://phorge.dev.yugabyte.com/D43101