[#26086] DocDB: Capture and return per RPC metrics for writes with protobuf response
Summary:
This diff captures the rocksdb and tablet metrics for writes and returns them
with the response. A future diff will use this to display these metrics in the
EXPLAIN (ANALYZE, DIST, DEBUG) output.

A single instance of `WriteQuery` can contain a batch of write operations.
If the `batch_tablet_metrics_update` gflag is true, docdb and tablet metrics
for all the write operations will be accumulated and stored in the `WriteQuery`
object. The accumulated metrics will be returned with the response PB of the
first write operation in the `WriteQuery` object.

Storing the metrics in the `WriteQuery` object ensures that we also capture
the tablet level metrics which are updated outside the scope of executing the
write operations.

We pass the metrics from the WriteQuery object to other places like
WaitOnConflictResolver while executing the query. In case of errors
like deadlock error, it's possible for the WriteQuery object to be destroyed
before WaitOnConflictResolver, in that case, it results in a use-after-free
bug. We solve this by using a shared pointer of the TabletMetrics pointer.
The destructor of the WriteQuery will update the TabletMetrics pointer to
the global TabletMetrics pointer. So, any metrics updated after the WriteQuery
object is destroyed will directly be updated in the global tablet metrics.
In this way, even after the WriteQuery object is destroyed, the shared pointer
is still alive until all the objects which store the metrics object are destroyed.

Perf report with this change

{F354351}

**Upgrade/rollback safety:**
This diff adds an optional metrics_capture field to PgsqlWriteRequestPB but does not require any flags to maintain upgrade/downgrade safety. If the field is not set, the default value (PGSQL_METRICS_CAPTURE_NONE) results in no metric changes being sent with the response/metric changes missing in the output, but it is acceptable for this feature to not report all metric changes during a rolling upgrade/downgrade, and the issue will resolve automatically when it finishes (all nodes sending a value after upgrade/value not checked at all after downgrade).
Jira: DB-15412

Test Plan: Jenkins

Reviewers: kramanathan, esheng

Reviewed By: esheng

Subscribers: svc_phabricator, yql, ybase

Differential Revision: https://phorge.dev.yugabyte.com/D41947