[BACKPORT 2024.2][#25710] YSQL: Fix path->rows set by yb_cost_bitmap_table_scan
Summary:
Remove the statements in `yb_cost_bitmap_table_scan` that overwrite `path->rows` with incorrect value based on only a subset of the predicates.

At the beginning of `yb_cost_bitmap_table_scan`, `path->rows` is set to the estimated number of output rows with selectivity of all the single relation predicate.  If any, join predicates that are parameterized and became single relation condition (for execution in a loop) are also taken into account.
```
    /* Mark the path with the correct row estimate */
    if (param_info)
        path->rows = param_info->ppi_rows;
    else
        path->rows = baserel->rows;
```

Then near the end of the function, `path->rows` is overwritten twice:
```
    path->rows =
        clamp_row_est(baserel->tuples *
                      clauselist_selectivity(root, baserel->baserestrictinfo,
                                             baserel->relid, JOIN_INNER, NULL));

    path->rows = tuples_fetched;
```
The problem with the first assignment is that `baserel->baserestrictinfo` does not include any parameterized join predicates.

The second assignment, where `tuples_fetched` is computed as follows, doesn't include the filters that are evaluated on the postgres side.
```
    tuples_fetched =
        clamp_row_est(adjusted_baserel_tuples *
                      clauselist_selectivity(root,
                                             list_union(indexquals,
                                                        rel_remote_quals),
                                             baserel->relid, JOIN_INNER, NULL));
```
Jira: DB-14968

Original commit: 03d9aaf29fe3ab584cdefacba094e79e504a3ae2 / D43109

- Adjusted the test for the optimizer's behavior difference.

Test Plan:
./yb_build.sh release --java-test 'org.yb.pgsql.TestPgRegressPlanner'  (also tried with `-n 100` to verify that the new test is not flakey)
./yb_build.sh release --java-test 'org.yb.pgsql.TestPgRegressYbBitmapScans'

Reviewers: telgersma

Reviewed By: telgersma

Subscribers: yql

Differential Revision: https://phorge.dev.yugabyte.com/D43143