[PLAT-17252]For re-configure YBC before backup/restore do it on all queryable nodes instead of just Running and ToBeRemoved ones
Summary:
We were only trying to re-configure YBC on nodes which are 'Running' or 'ToBeRemoved'. But
since backup can be run aftre a failed tsk, there are nodes which may be in other states. Changed
the nodes filter to `Queryable` nodes.

Also, during YBC ping test, there was a rare case when we may not be closing the client for the YBC
server that we do not use. Fixed that case.

Test Plan: pipeline

Reviewers: nsingh

Reviewed By: nsingh

Differential Revision: https://phorge.dev.yugabyte.com/D43060