[#26782] YSQL: Remove redundant yb.orig.planner_taqo_basic test
Summary:
This test is generated from the `basic` workload in TAQO. We will remove
this test because it is redundant and hard to maintain. These tests were
commented out and not in use recently. This was one of the first
workloads we designed for TAQO, and we plan to remove this from
TAQO tests too. We will keep other tests derived from workloads
`seek-next-estimation` and `tuning-tests` as these are more relevant.
Jira: DB-16162

Test Plan: ./yb_build.sh --java-test 'org.yb.pgsql.TestPgRegressTAQO'

Reviewers: jason, mihnea

Reviewed By: jason

Subscribers: yql

Differential Revision: https://phorge.dev.yugabyte.com/D43171