[#26624] yugabyted: Fix yugabyted-ui VCPU counts for k8s deployments
Summary:
Currently in yugabyted-ui for k8s deployments, the VCPU and total core counts shown in the UI correspond to the total cores/vcpu of the underlying VM.
Instead, the UI will be changed to not count the number of cores/VCPU in cases of k8s deployments.
This will be done by checking an environment variable YUGABYTED_UI_K8S which will be set in the yugabyte-db helm chart.
Jira: DB-16000

Test Plan: manual tests

Reviewers: nikhil

Reviewed By: nikhil

Subscribers: yugabyted-dev, djiang

Differential Revision: https://phorge.dev.yugabyte.com/D43181