[#26702] DocDB: Add BSON type to yb
Summary:
Adding BSON as a dedicated data type. This is required for the Microsoft DocumentDB extension integration.
The extension was imported as part of b01e29bafa6fb8bfb899f0b3ac6e98363340c8e7/D43110.

The type currently maps to binary data type since we don't have any index support for this type yet. Mapping it to bson library and handling conversion to comparable binary format for the RocksDB key will be done in a later diff.

**Upgrade/Rollback safety**
The new type is not used in any existing code. The feature will be guarded by a AutoFlag when the DocumentDB extension is made available.

Fixes#26702
Jira: DB-16078

Test Plan:
build --clang17 --cxx-test pggate_test_type

Regress tests will be added after the extension is merged with the new OID from this change.

Reviewers: slingam, sergei, tnayak

Reviewed By: sergei, tnayak

Subscribers: yql, ybase

Differential Revision: https://phorge.dev.yugabyte.com/D43048