[#26717] YSQL: Index Recheck Correctness fix
Summary:
=== Background ===
In Postgres and Yugabyte, sometimes an Index Scan may "require recheck" - the rows returned from the index may not be fully accurate, and so the expression must be re-evaluated locally to remove any rows that do not match. In Yugabyte, these cases are:
* when multiple conditions are bound to the same column (e.g. `a IN (1,2,3,6) AND a <= 3`)
* cross type operators (e.g. comparing an int4 to an int8)
* Some RowComparisonExpressions
* Incomplete bindings to hash partition keys
* queries that rely on the `ASC` column of an index created on `(col1 HASH, col2 ASC)`. See issue [[ https://github.com/yugabyte/yugabyte-db/issues/26742 | #26742 ]]. **This one is key to this issue**

If the values required by the recheck are not fetched, then the recheck will run incorrectly and remove the row.

=== The issue ===
There are several cases in each of the referenced github issues of an index on `(col1 HASH, col1 ASC)` where the query does not specifically target these columns. It uses the columns to filter, but the values themselves are not sent back to the client node.

In PG15 commit 9a3ddeb519e8d91b670fa923619b212a9670389c, the expression used to rechecking was switched to a special `recheckqual` rather than the original `indexqual`. The motivation for this change was to address heap tuple visibility, which is not a concern in Yugabyte.  `recheckqual` goes through different transformations, and most notably, refers to the table attributes rather than the index attributes. Yugabyte has its own code to deal with selecting targets, and this was not updated to reflect this PG change.

=== The fix ===
Update `ybcBuildRequiredAttrs` to account for the attributes mentioned in the recheck expression. Since these expressions have different attribute numbers, it may result in the same column being requested from DocDB multiple times (once as the first column of the index, and a second time as the second column). This is only an issue where a column appears multiple times in an index definition. For example, in the test
```lang=sql
set yb_fetch_row_limit = 0;
set yb_fetch_size_limit = '3kB';
insert into simple select * from generate_series(2, 1000);
explain (analyze, dist, costs off, timing off, summary off)
/*+ IndexOnlyScan(simple) */ SELECT 'a' FROM simple WHERE i < 5;
                               QUERY PLAN
-------------------------------------------------------------------------
 Index Only Scan using simple_i_i1_idx on simple (actual rows=4 loops=1)
   Index Cond: (i < 5)
   Rows Removed by Index Recheck: 996
   Heap Fetches: 0
   Storage Index Read Requests: 6
   Storage Index Rows Scanned: 1003
(6 rows)
```

There are six index read requests to read the whole table. Whereas in the plan before this fix:
```lang=sql
/*+ IndexOnlyScan(simple) */ explain (analyze, dist, costs off, timing off, summary off) SELECT 'a' FROM simple WHERE i<5;
                               QUERY PLAN
-------------------------------------------------------------------------
 Index Only Scan using simple_i_i1_idx on simple (actual rows=0 loops=1)
   Index Cond: (i < 5)
   Rows Removed by Index Recheck: 1000
   Heap Fetches: 0
   Storage Index Read Requests: 2
   Storage Index Rows Scanned: 1000
(6 rows)
```
the results are incorrect, but it takes less requests to scan the table because only one column is fetched instead of two. Ideally we could identify that although the column_refs are different (1, 2), they refer to the same physical column and so aren't necessary to differentiate between. However, this only applies to index only scans where a column is repeated in the index and recheck is required, so the impact is limited.

Without this fix,

Also, `yb_indexqual_for_recheck` was a copy of `indexqual` with some `yb_hash_code` functions modified. Since PG has now introduced a `recheckqual` that undergoes different transformations than `indexqual`, this field is no longer necessary. This means that all functions associated with creating this field can be removed as well.
Jira: DB-16091

Test Plan:
```
./yb_build.sh --java-test TestPgRegressIndex
./yb_build.sh --java-test TestPgRegressProc
```

Reviewers: tnayak, amartsinchyk

Reviewed By: amartsinchyk

Subscribers: yql

Differential Revision: https://phorge.dev.yugabyte.com/D43082