[BACKPORT 2024.2][#26788] docdb: Use geographically closest tserver for clone
Summary:
Clone runs ysql_dump on a tserver, which performs a lot of queries against pg_catalog, which is on yb-master. This can take a long time if the tserver is not in the same region / zone as the master. This causes the clone to time out after 10 minutes. After this diff, the clone runs in a couple seconds. Currently we pick a random tserver, but we should pick the one closest to master.

This diff also moves the clone tasks to their own file.
Jira: DB-16169

Original commit: 42258a162c0089b7d8a049fabc330e2a6772c37c / D43177

Test Plan:
There are already tests for GetClosestLiveTserver in master-test.cc. I've enhanced one of the clone mock tests to ensure that we actually pass the returned tserver to
SchedulePgCloneSchemaTask.

Also test on a YBA multi-region cluster to verify that the clone finishes quickly instead of timing out (as it did on the same cluster before upgrading).

Reviewers: mhaddad

Reviewed By: mhaddad

Subscribers: ybase

Differential Revision: https://phorge.dev.yugabyte.com/D43227