[#25500] YSQL: Introduce 'server_type' option to differentiate postgres_fdw servers
Summary:
postgres_fdw supports any database that uses the postgres communication protocol as a foreign server.
For yugabyteDB, this implies that a configured foreign server could be one of the following:
 - A PostgreSQL instance
 - A different YugabyteDB cluster
 - The same YugabyteDB cluster

Although postgres and yugabyteDB share a common communication protocol, there are important distinctions such as:
 - Postgres uses the system column `ctid` to uniquely identify a tuple while yugabyteDB uses the computed ybctid

This creates a requirement to identify the type of underlying server (yugabyteDB, postgres or other) to ensure compatibility.
This revision introduces a new `server_type` option to the CREATE and ALTER SERVER commands to differentiate between server types.
This option is written to the column `srvoptions` of `pg_foreign_server` where the FDW is of type `postgres_fdw`.

**New Foreign Server Semantics**
`server_type` accepts the three case-insensitive values `["postgreSQL, "yugabyteDB", <empty/unset>]` with the following semantics:
 - `postgreSQL` indicates that the underlying server is of type postgres.
 - `yugabyteDB` indicates that an underlying yugabyteDB cluster.
 - If left unset, the underlying server defaults to postgres.
 - Once set, modifying or dropping server types is disallowed.

**Upgrade Semantics**
Since `server_type` is written to an existing column in an existing catalog table, no special upgrade handling or migration scripts are required.
This revision assumes that existing users of postgres_fdw extension in yugabyteDB currently make use of only postgres as a foreign server.
The assumption is realistic as support for certain key features such as use_remote_estimate, non-pushed down UPDATEs, DELETEs is currently missing with yugabyetDB as a foreign server.
When such users of postgres upgrade, `server_type` will remain unpopulated for the foreign server entry.
However, since we default to using postgres when `server_type` is unset, this change will be transparent to them, and their operations will remain unaffected.

Existing users of yugabyteDB as a foreign server will be advised to manually set the `server_type` of their foreign servers via the command:
```
ALTER SERVER <name> OPTIONS (server_type 'yugabyteDB');
```
Jira: DB-14759

Test Plan:
Run the following tests:
```
./yb_build.sh --java-test 'org.yb.pgsql.TestPgRegressContribYbPostgresFdw#schedule'
./yb_build.sh --java-test 'org.yb.pgsql.TestPgRegressContribPostgresFdw#schedule'
```

Reviewers: fizaa, smishra, #db-approvers

Reviewed By: fizaa, #db-approvers

Subscribers: rthallam, svc_phabricator, jason, yql

Tags: #jenkins-ready

Differential Revision: https://phorge.dev.yugabyte.com/D43119